changes in env initialisation
