+++ backend/.env
@@ -1,2 +1,2 @@
-GROQ_API_KEY=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
-MONGODB_URI=http://localhost:27017/resume_ranking
+GROQ_API_KEY=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
+MONGODB_URI=mongodb+srv://rakheshkrishnap:lpcz6GDVvAfud8XM@newsspyder.no9qeon.mongodb.net/resume_ranking?retryWrites=true&w=majority&appName=NewsSpyder
