+++ backend/utils/pdf_processor.py
@@ -1,0 +1,6 @@
+import fitz
+
+def extract_text_from_pdf(pdf_path):
+    doc = fitz.open(pdf_path)
+    return "\n".join(page.get_text() for page in doc)
+
